requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
WTForms
Pillow
python-barcode
pyserial